refactor(config): 🔨 ファイル配置を変更
