+++ packages/schema/src/prefecture/index.ts
@@ -1,2 +1,1 @@
 export * from './prefecture.schema';
-export * from './prefecture.enum';
