+++ packages/schema/src/prefecture/stats/index.ts
@@ -1,1 +1,3 @@
 export * from './prefecture-stats.schema';
+export * from './prefecture-stats.enum';
+export * from './prefecture-stats.config';
