+++ packages/schema/src/prefecture/stats/prefecture-stats.enum.ts
@@ -1,0 +1,3 @@
+export const PrefectureStatsName = ['POPULATION', 'AREA'] as const;
+
+export type PrefectureStatsName = (typeof PrefectureStatsName)[number];
