+++ packages/schema/src/todofuken/game/index.ts
@@ -1,2 +1,3 @@
 export * from './game.schema';
 export * from './game.enum';
+export * from './game-difficulty.config';
